+++ mkinit/formatting.py
@@ -34,6 +34,8 @@ def _ensure_options(given_options=None):
         'with_mods': True,
         'with_all': True,
         'relative': False,
+        'lazy_import': False,
+        'lazy_boilerplate': None,
     }
     options = default_options.copy()
     for k in given_options.keys():
@@ -210,6 +212,13 @@ def _initstr(modname, imports, from_imports, explicit=set(), protected=set(),
         python -m mkinit.static_autogen _initstr
 
     Args:
+        modname (str): the name of the module to generate the init str for
+
+        imports (List[str]): list of module-level imports
+
+        from_imports (List[Tuple[str, List[str]]]):
+            List of submodules and their imported attributes
+
         options (dict): customize output
 
     CommandLine:
@@ -243,6 +252,37 @@ def _initstr(modname, imports, from_imports, explicit=set(), protected=set(),
         __all__ = ['a', 'b', 'bar', 'baz', 'c', 'd', 'e', 'f', 'g', 'h', 'i', 'j', 'k',
                    'l', 'm', 'n', 'o', 'p', 'q', 'r', 's', 't', 'u', 'v', 'w', 'x',
                    'y', 'z']
+
+    Example:
+        >>> modname = 'foo'
+        >>> imports = ['.bar', '.baz']
+        >>> from_imports = [('.bar', ['func1', 'func2'])]
+        >>> options = {'lazy_import': 1, 'lazy_boilerplate': None}
+        >>> initstr = _initstr(modname, imports, from_imports, options=options)
+        >>> print(initstr)
+        ...
+
+        >>> options = {'lazy_import': 1, 'lazy_boilerplate': 'from importlib import lazy_import'}
+        >>> initstr = _initstr(modname, imports, from_imports, options=options)
+        >>> print(initstr.replace('\n\n', '\n'))
+        from importlib import lazy_import
+        __getattr__ = lazy_install(
+            __name__,
+            submodules={
+                'bar',
+                'baz',
+            },
+            submod_attrs={
+                'bar': [
+                    'func1',
+                    'func2',
+                ],
+            },
+        )
+        def __dir__():
+            return __all__
+        __all__ = ['bar', 'baz', 'func1', 'func2']
+
     """
     options = _ensure_options(options)
 
@@ -254,18 +294,8 @@ def _initstr(modname, imports, from_imports, explicit=set(), protected=set(),
     # if options.get('with_header', False):
     #     parts.append(_make_module_header())
 
-    def append_part(new_part):
-        """ appends a new part if it is nonempty """
-        if new_part:
-            if parts:
-                # separate from previous parts with a newline
-                parts.append('')
-            parts.append(new_part)
-
     if options.get('with_mods', True):
         explicit_exports.extend([e.lstrip('.') for e in imports])
-        append_part(_make_imports_str(imports, modname))
-
     if options.get('with_attrs', True):
         from fnmatch import fnmatch
         # TODO: allow pattern matching here
@@ -281,14 +311,14 @@ def _initstr(modname, imports, from_imports, explicit=set(), protected=set(),
         _pp_pats = protected_pats | private_pats
         _pp_set = private_set | protected_set
 
+        def _private_matches(x):
+            x = x.lstrip('.')
+            return x in private_set or any(fnmatch(x, pat) for pat in private_pats)
+
         def _pp_matches(x):
             # TODO: standardize how explicit vs submodules are handled
             x = x.lstrip('.')
             return x in _pp_set or any(fnmatch(x, pat) for pat in _pp_pats)
-
-        def _private_matches(x):
-            x = x.lstrip('.')
-            return x in private_set or any(fnmatch(x, pat) for pat in private_pats)
 
         _from_imports = [
             (m, sub) for m, sub in from_imports if not _pp_matches(m)
@@ -298,10 +328,125 @@ def _initstr(modname, imports, from_imports, explicit=set(), protected=set(),
             n for m, sub in _from_imports for n in sub
             if not _private_matches(n)
         ])
-        attr_part = _make_fromimport_str(_from_imports, modname)
-        append_part(attr_part)
+
+    def append_part(new_part):
+        """ appends a new part if it is nonempty """
+        if new_part:
+            if parts:
+                # separate from previous parts with a newline
+                parts.append('')
+            parts.append(new_part)
+
+    if options['lazy_import']:
+        import ubelt as ub
+
+        default_lazy_boilerplate = ub.codeblock(
+            r'''
+
+            DEBUG_LAZY = 0
+
+
+            def lazy_install(module_name, submodules, submod_attrs):
+                """
+                Defines gettr for lazy import via PEP 562
+                https://www.python.org/dev/peps/pep-0562/
+                """
+                import sys
+                import importlib
+                import importlib.util
+                all_funcs = []
+                for mod, funcs in submod_attrs.items():
+                    all_funcs.extend(funcs)
+                name_to_submod = {
+                    func: mod for mod, funcs in submod_attrs.items()
+                    for func in funcs
+                }
+
+                def require(fullname):
+                    if fullname in sys.modules:
+                        return sys.modules[fullname]
+
+                    spec = importlib.util.find_spec(fullname)
+                    try:
+                        module = importlib.util.module_from_spec(spec)
+                    except:
+                        raise ImportError(f'Could not lazy import module {fullname}') from None
+                    loader = importlib.util.LazyLoader(spec.loader)
+
+                    sys.modules[fullname] = module
+
+                    # Make module with proper locking and get it inserted into sys.modules.
+                    loader.exec_module(module)
+
+                    return module
+
+                def __getattr__(name):
+                    if DEBUG_LAZY:
+                        print('LAZY EVALUATE {!r} in {}'.format(name, __name__))
+                    if name in submodules:
+                        fullname = f'{module_name}.{name}'
+                        attr = require(fullname)
+                    elif name in name_to_submod:
+                        modname = name_to_submod[name]
+                        module = importlib.import_module(
+                            f'{module_name}.{modname}'
+                        )
+                        attr = getattr(module, name)
+                    else:
+                        raise AttributeError(f'No {module_name} attribute {name}')
+                    if DEBUG_LAZY:
+                        print('EVALUATED TO = {!r}'.format(attr))
+                    # Set module-level attribute so getattr is not called again
+                    globals()[name] = attr
+                    return attr
+                return __getattr__
+            '''
+        )
+        template = ub.codeblock(
+            '''
+            __getattr__ = lazy_install(
+                __name__,
+                submodules={submodules},
+                submod_attrs={submod_attrs},
+            )
+            ''')
+        submod_attrs = {}
+        for submod, attrs in from_imports:
+            submod = submod.lstrip('.')
+            submod_attrs[submod] = attrs
+        submodules = {m.lstrip('.') for m in imports}
+        initstr = template.format(
+            submodules=ub.repr2(submodules),
+            submod_attrs=ub.repr2(submod_attrs)
+        )
+
+        import black
+        initstr = black.format_str(
+            initstr, mode=black.Mode(string_normalization=False))
+
+        if options['lazy_boilerplate'] is None:
+            append_part(default_lazy_boilerplate)
+        else:
+            # Customize lazy boilerplate
+            append_part(options['lazy_boilerplate'])
+
+        append_part(initstr.rstrip())
+    else:
+        if options.get('with_mods', True):
+            append_part(_make_imports_str(imports, modname))
+
+        if options.get('with_attrs', True):
+
+            attr_part = _make_fromimport_str(_from_imports, modname)
+            append_part(attr_part)
 
     if options.get('with_all', True):
+        if options['lazy_import']:
+            append_part(ub.codeblock(
+                '''
+                def __dir__():
+                    return __all__
+                '''))
         exports_repr = ["'{}'".format(e)
                         for e in sorted(explicit_exports)]
         rhs_body = ', '.join(exports_repr)
@@ -365,8 +510,6 @@ def _packed_rhs_text(lhs_text, rhs_text):
         >>> rhs_text = ', '.join(fromlist) + ',)'
         >>> packstr = _packed_rhs_text(lhs_text, rhs_text)
         >>> print(packstr)
-
-
     """
     # FIXME: the parens get broken up wrong
     # filler = '-' * (len(lhs_text) - 1) + ' '
@@ -384,6 +527,7 @@ def _packed_rhs_text(lhs_text, rhs_text):
         # not sure why this isn't 76? >= maybe?
         max_width = 79
 
+        # This is a hacky heuristic that could perhaps be more robust?
         if len(lhs_text) > max_width * 0.7:
             newline_prefix = ' ' * 4
         else:
@@ -393,7 +537,6 @@ def _packed_rhs_text(lhs_text, rhs_text):
         wrapped_lines = textwrap.wrap(
             raw_text,
             break_long_words=False,
-
             width=79, initial_indent='',
             subsequent_indent=newline_prefix)
         packstr = '\n'.join(wrapped_lines)