+++ tests/test_with_dummy.py
@@ -2,12 +2,16 @@
 Tests run on a dummy package
 """
 import ubelt as ub
+from distutils.version import LooseVersion
+import sys
 from os.path import join
 
 
-def make_dummy_package(dpath, pkgname='mkinit_dummy_module'):
+def make_dummy_package(dpath, pkgname='mkinit_dummy_module', side_effects=0):
     """
     Creates a dummy package structure with or without __init__ files
+
+    ANY EXISTING FILES ARE DELETED
     """
     root = ub.ensuredir(join(dpath, pkgname))
     ub.delete(root)
@@ -198,24 +202,30 @@ def make_dummy_package(dpath, pkgname='mkinit_dummy_module'):
     # Ensure that each submodule has an __init__
     # (do we need this with PEP 420 anymore?)
     root = paths['root']
-    from os.path import relpath, exists
+    from os.path import relpath, exists, dirname
     import os
+
+    dpath = dirname(root)
+    init_fpaths = set()
     for key, path in paths.items():
-        relative = relpath(path, root)
+        relative = relpath(path, dpath)
         suffix = []
         parts = relative.split(os.sep)
         for part in parts:
             if '.' not in part:
                 suffix.append(part)
-                middir = join(root, os.sep.join(suffix))
+                middir = join(dpath, os.sep.join(suffix))
                 fpath = join(middir, '__init__.py')
-                if not exists(fpath):
-                    ub.touch(fpath)
+                init_fpaths.add(fpath)
+
+    for fpath in init_fpaths:
+        if not exists(fpath):
+            ub.touch(fpath)
 
     ub.writeto(paths['long_submod'], ub.codeblock(
         '''
         def a_very_nested_function():
-            pass
+            print('a_very_nested_function in {}'.format(__file__))
         def another_func1():
             pass
         def another_func2():
@@ -226,6 +236,14 @@ def make_dummy_package(dpath, pkgname='mkinit_dummy_module'):
             pass
         '''
     ))
+
+    if side_effects:
+        with open(paths['long_submod'], 'a') as file:
+            file.write(ub.codeblock(
+                '''
+                print('NESTED SIDE EFFECT')
+                '''
+            ))
     return paths
 
 
@@ -300,6 +318,79 @@ def test_dynamic_init():
     for i in range(1, 15):
         want = 'good_attr_{:02d}'.format(i)
         assert want in text, 'missing {}'.format(want)
+
+
+def test_lazy_import():
+    """
+    python ~/code/mkinit/tests/test_with_dummy.py test_lazy_import
+    """
+    import pytest
+    import mkinit
+
+    cache_dpath = ub.ensure_app_cache_dir('mkinit/tests')
+    paths = make_dummy_package(cache_dpath)
+
+    pkg_path = paths['root']
+    mkinit.autogen_init(pkg_path, options={'lazy_import': 1}, dry=False, recursive=True)
+
+    if LooseVersion('{}.{}'.format(*sys.version_info[0:2])) < LooseVersion('3.7'):
+        pytest.skip()
+
+    from os.path import dirname
+    dpath = dirname(paths['root'])
+    with ub.util_import.PythonPathContext(dpath):
+        import mkinit_dummy_module
+        print('mkinit_dummy_module = {!r}'.format(mkinit_dummy_module))
+        print(dir(mkinit_dummy_module))
+        print('mkinit_dummy_module.a_very_nested_function = {!r}'.format(mkinit_dummy_module.a_very_nested_function))
+        mkinit_dummy_module.a_very_nested_function()
+
+
+def test_recursive_lazy_autogen():
+    """
+    xdoctest ~/code/mkinit/tests/test_with_dummy.py test_recursive_lazy_autogen
+    """
+    import pytest
+    import mkinit
+    cache_dpath = ub.ensure_app_cache_dir('mkinit/tests')
+    paths = make_dummy_package(cache_dpath, pkgname='mkinit_rec_lazy_autogen', side_effects=True)
+    pkg_path = paths['root']
+
+    mkinit.autogen_init(pkg_path, options={'lazy_import': 1}, dry=False, recursive=True)
+
+    if LooseVersion('{}.{}'.format(*sys.version_info[0:2])) < LooseVersion('3.7'):
+        pytest.skip()
+
+    with ub.util_import.PythonPathContext(cache_dpath):
+        import mkinit_rec_lazy_autogen
+        print('mkinit_rec_lazy_autogen = {!r}'.format(mkinit_rec_lazy_autogen))
+        print('mkinit_rec_lazy_autogen.good_attr_01 = {!r}'.format(mkinit_rec_lazy_autogen.good_attr_01))
+        print('mkinit_rec_lazy_autogen.a_very_nested_function = {!r}'.format(mkinit_rec_lazy_autogen.a_very_nested_function))
+        print('mkinit_rec_lazy_autogen.a_very_nested_function = {!r}'.format(mkinit_rec_lazy_autogen.a_very_nested_function))
+        print('mkinit_rec_lazy_autogen.a_very_nested_function = {!r}'.format(mkinit_rec_lazy_autogen.a_very_nested_function))
+        mkinit_rec_lazy_autogen.a_very_nested_function()
+
+
+def test_recursive_eager_autogen():
+    """
+    xdoctest ~/code/mkinit/tests/test_with_dummy.py test_recursive_eager_autogen
+    """
+    import mkinit
+    cache_dpath = ub.ensure_app_cache_dir('mkinit/tests')
+    paths = make_dummy_package(cache_dpath, pkgname='mkinit_rec_eager_autogen', side_effects=True)
+    pkg_path = paths['root']
+
+    mkinit.autogen_init(pkg_path, options={'lazy_import': 0}, dry=False, recursive=True)
+
+    with ub.util_import.PythonPathContext(cache_dpath):
+        import mkinit_rec_eager_autogen
+        print('mkinit_rec_eager_autogen = {!r}'.format(mkinit_rec_eager_autogen))
+        print('mkinit_rec_eager_autogen = {!r}'.format(mkinit_rec_eager_autogen))
+        print('mkinit_rec_eager_autogen.good_attr_01 = {!r}'.format(mkinit_rec_eager_autogen.good_attr_01))
+        print('mkinit_rec_eager_autogen.a_very_nested_function = {!r}'.format(mkinit_rec_eager_autogen.a_very_nested_function))
+        print('mkinit_rec_eager_autogen.a_very_nested_function = {!r}'.format(mkinit_rec_eager_autogen.a_very_nested_function))
+        print('mkinit_rec_eager_autogen.a_very_nested_function = {!r}'.format(mkinit_rec_eager_autogen.a_very_nested_function))
+        mkinit_rec_eager_autogen.a_very_nested_function()
 
 
 if __name__ == '__main__':